Move build functions to the build submodule and call it using pear-ipc
